backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-dotenv==1.0.0
